Updated packages, added component folders, renamed 'Concerts' to 'ConcertHistory'
