+++ package.json
@@ -13,8 +13,8 @@
     "@testing-library/user-event": "^13.5.0",
     "googleapis": "^100.0.0",
     "moment-timezone": "^0.5.34",
-    "react": "^18.0.0",
-    "react-dom": "^18.0.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-google-calendar-api": "^1.5.2",
     "react-scripts": "^5.0.1",
     "web-vitals": "^2.1.4"
@@ -44,7 +44,7 @@
     ]
   },
   "devDependencies": {
-    "react-router-dom": "^6.3.0",
-    "@babel/plugin-proposal-private-property-in-object":"7.21.11"
+    "@babel/plugin-proposal-private-property-in-object": "7.21.11",
+    "react-router-dom": "^6.3.0"
   }
 }
