+++ src/index.js
@@ -2,8 +2,8 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import './index.css';
-import App from './App';
-import Concerts from './Concerts'
+import App from './App/App';
+import ConcertHistory from './ConcertHistory/ConcertHistory';
 import reportWebVitals from './reportWebVitals';
 
 export default function Index() {
@@ -11,7 +11,7 @@ export default function Index() {
     <BrowserRouter>
       <Routes>
           <Route index element={<App />} />
-          <Route path="concerts" element={<Concerts />} />
+          <Route path="concerts" element={<ConcertHistory />} />
       </Routes>
     </BrowserRouter>
   );
